feat: 이미지 압축 기능 추가
