+++ app/recoil/atoms.ts
@@ -12,18 +12,3 @@ export const editorState = atom<string>({
   key: 'editorState',
   default: '',
 });
-
-export const imageState = atom<string>({
-  key: 'imageState',
-  default: '',
-});
-
-export const imageUploadState = atom<File | null>({
-  key: 'imageUploadState',
-  default: null,
-});
-
-export const uploadPreviewState = atom<string[]>({
-  key: 'uploadPreviewState',
-  default: [],
-});
